+++ yash-env/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `impl yash_system::alias::Glossary for Env`
+- `input::Echo`
+    - This is a decorator of `Input` that implements the behavior of the verbose shell option.
 - `input::FdReader` is now marked `#[must_use]`.
 
 ### Changed